requirements.txt
absl-py==0.9.0
astor==0.6.2
backports.weakref==1.0rc1
bleach==1.5.0
certifi==2019.11.28
cloudpickle==1.3.0
colorama==0.4.3
cycler==0.10.0
Cython==0.25.2
dask==2.10.1
decorator==4.4.1
entrypoints==0.2.3
enum34==1.1.6
futures==3.1.1
gast==0.2.2
google-pasta==0.1.8
grpcio==1.10.0
h5py==2.7.1
html5lib==0.9999999
imageio==2.6.1
imgaug==0.2.7
imutils==0.5.3
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
jedi==0.10.2
Jinja2==2.9.6
joblib==0.14.1
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==5.1.0
jupyter-core==4.6.3
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==2.6.8
MarkupSafe==1.0
matplotlib==3.1.2
mistune==0.8.1
nbconvert==5.2.1
nbformat==4.3.0
networkx==2.4
nose==1.3.7
notebook==5.7.8
numpy==1.18.1
opencv-contrib-python==4.2.0.32
opencv-python==4.1.2.30
opencv-python-headless==4.1.2.30
opt-einsum==3.1.0
pandas==0.20.2
pandocfilters==1.4.1
pexpect==4.2.0
pickleshare==0.7.4
Pillow==7.0.0
prometheus-client==0.7.1
prompt-toolkit==1.0.14
protobuf==3.6.1
ptyprocess==0.5.2
pydot==1.2.3
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz===2011k
PyWavelets==1.1.1
pywin32==227
pywinpty==0.5.7
PyYAML==5.1
pyzmq==18.1.1
qtconsole==4.3.0
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.0.0
Send2Trash==1.5.0
Shapely==1.7.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
tensorboard==1.15.0
tensorflow==1.15.2
tensorflow-estimator==1.15.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.3.1
toolz==0.10.0
tornado==4.5.1
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.3
widgetsnbextension==2.0.0
wincertstore==0.2
wrapt==1.12.0